requirements.txt
flask==3.0.3
pillow==10.3.0
